move import to characters cog
